requirements.txt
fastapi
uvicorn
pydantic

streamlit

langchain
langchain-community
langchain-openai
faiss-cpu
rank-bm25

sentence-transformers

loguru
pypdf
databricks-sql-connector
python-dotenv